feat: Allow a user to subscribe to notifications on an application
Closes gravitee-io/issues#3115
